requirements.txt
Bootstrap_Flask==2.2.0
Flask~=3.0.2
Flask_CKEditor==0.4.6
Flask_Login==0.6.2
Flask-Gravatar==0.5.0
flask_sqlalchemy==3.0.5
Flask_WTF==1.1.1
Werkzeug~=3.0.1
WTForms==3.0.1
SQLAlchemy==2.0.19
gunicorn==21.2.0
psycopg2-binary==2.9.6
mailjet-rest==1.3.4
requests~=2.31.0
twilio==9.0.1
PyPDF2~=2.12.1
pillow~=10.2.0
colorthief~=0.2.1